Edit css file to style the elements in the body
In style.css:

Add background-color of pink and set font-family
of odd lines, style second line to have blue text
and font-size of 36px, set font-size to 24px for
third line, set background-color to light green
with font-size of 24px and font-weight of bold
to fourth line.

In index.html:

Add appropriate class and id names to elements to
use styles added above.